+++ src/types/api.ts
@@ -1,0 +1,8 @@
+import { Book } from "./book";
+
+export type Status = { end: number; total: number };
+
+export type FetchResults = {
+  books: Book[];
+  status: Status;
+};
