+++ src/types/index.ts
@@ -1,2 +1,3 @@
+export * from "./api";
 export * from "./book";
 export * from "./props";
