+++ src/util/html-utils.ts
@@ -1,56 +1,4 @@
-import { Book, Props } from "../types";
-import { getUrl } from "./get-url";
-
-type Status = { end: number; total: number };
-
-export type FetchResults = {
-  books: Book[];
-  status: Status;
-};
-
-export const fetchPage = async (page: number, props: Props): Promise<FetchResults> => {
-  // Get the books from Goodreads
-  const url = getUrl(props, page);
-  url.searchParams.append("page", String(page));
-  const response = await window.fetch(url.toString(), { headers: { accept: "text/javascript" } });
-
-  // Simulate success if we get a 204 No Content response
-  if (response.status === 204) {
-    return {
-      books: [],
-      status: {
-        end: page * 30,
-        total: 0
-      }
-    };
-  }
-
-  const responseBody = await response.text();
-  const { html, status } = parseJsonP(responseBody);
-  const table = `<table>${html}</table>`;
-  const books = getBooksFromHtml(table, props.width);
-  return {
-    books,
-    status
-  };
-};
-
-const parseJsonP = (jsonp: string): { html: string; status: Status } => {
-  const [html, status] = jsonp.split("\n");
-
-  // eslint-disable-next-line quotes
-  const json = html.replace('Element.insert("booksBody", ', "").replace(" });", "}").replace("bottom", '"bottom"');
-  const output: string = JSON.parse(json).bottom;
-
-  const matches = status.match(/(?<end>\d*) of (?<total>\d*) loaded/);
-  return {
-    html: output,
-    status: {
-      end: parseInt(matches?.groups?.end ?? "0"),
-      total: parseInt(matches?.groups?.total ?? "0")
-    }
-  };
-};
+import { Book } from "../types";
 
 export const getBooksFromHtml = (html: string, width: string | number | undefined = 150): Book[] => {
   const parser = new DOMParser();
