+++ src/util/index.ts
@@ -1,3 +1,3 @@
+export * from "./api";
 export * from "./get-url";
 export * from "./html-utils";
-export * from "./xml2js-utils";
